requirements.txt
numpy 
pandas
psycopg2
requests
yfinance
scikit-learn
python-dotenv
fredapi
alpaca-py
bs4
python-dotenv